Add export flat to st.plot in dashboards to allow pulling them into edit workflow easily
